remove record
