+++ anchor/network/src/handshake/record/mod.rs
@@ -1,3 +1,2 @@
 pub mod envelope;
-pub mod record;
 pub mod signing;
